Public

Public interfaces

Documentation for SpinMonteCarlo.jl's public interface (exported).

Driver

SpinMonteCarlo.runMC — Function.

runMC(param::Parameter)
runMC(params::AbstractArray{Parameter} ; parallel::Bool=false)

Runs Monte Carlo simulation(s) and returns calculated observables.

Keyward aruguments

parallel: If true, runs simulations in parallel (uses pmap instead of map).

Required keys in param

"Model"

"Update Method"

Optional keys in param

"MCS": The number of Monte Carlo steps after thermalization

Default: 8192

"Thermalization": The number of Monte Carlo steps for thermalization

Default: MCS>>3

"Seed": The initial seed of the random number generator, MersenneTwister

Default: determined randomly (see Random.seed!)

source

Model

SpinMonteCarlo.Ising — Type.

Ising model with energy $E = -\sum_{ij} J_{ij} \sigma_i \sigma_j$, where $\sigma_i$ takes value of 1 (up spin) or -1 (down spin).

source

SpinMonteCarlo.Potts — Type.

Q state Potts model with energy $E = -\sum_{i,j} \delta_{\sigma_i, \sigma_j}$, where $\sigma_i$ takes an integer value from $1$ to $Q$ and $\delta$ is a Kronecker's delta. Order parameter (total magnetization) is defined as \begin{equation} M = \frac{Q-1}{Q}N1 - \frac{1}{Q}(N-N1), \end{equation} where $N$ is the number of sites and $N_1$ is the number of $\sigma=1$ spins.

source

SpinMonteCarlo.Clock — Type.

Q state clock model with energy $E = -\sum_{ij} J_{ij} \cos(\theta_i - \theta_j)$, where $\theta_i = 2\pi \sigma_i/Q$ and $\sigma_i$ takes an integer value from $1$ to $Q$.

source

SpinMonteCarlo.XY — Type.

XY model with energy $E = -\sum_{ij} J_{ij} \cos(\theta_i - \theta_j)$, where $\theta_i = 2\pi \sigma_i$ and $\sigma_i \in [0, 1)$.

source

SpinMonteCarlo.QuantumXXZ — Type.

Spin-$S$ XXZ model represented as the following Hamiltonian, \begin{equation} \mathcal{H} = \sum{i,j} \left[ J{ij}^z Si^z Sj^z + \frac{J{ij}^{xy}}{2} (Si^+ Sj^- + Si^-Sj^+) \right] - \sumi \Gammai Si^x, \end{equation} where $S^x, S^y, S^z$ are $x, y$ and $z$ component of spin operator with length $S$, and $S^\pm \equiv S^x \pm iS^y$ are ladder operator. A state is represented by a product state (spins at $\tau=0$) of local $S^z$ diagonal basis and an operator string (perturbations). A local spin with length $S$ is represented by a symmetrical summation of $2S$ sub spins with length $1/2$.

source

Lattice generator

SpinMonteCarlo.dimer_lattice — Function.

dimer_lattice()
dimer_lattice(param::Dict)

Generates dimer lattice (indeed, this is not a "lattice")

source

SpinMonteCarlo.chain_lattice — Function.

chain_lattice(L::Integer)
chain_lattice(param::Dict)

Generates chain lattice with length L = param["L"]

nsitetypes is 2: all 1 (2) sites do not connect to another 1 (2) site.

nbondtypes is 2: 1 bond connects 2n-1 and 2n sites and 2 bond connects 2n and 2n+1 sites.

source

SpinMonteCarlo.square_lattice — Function.

square_lattice(L::Integer, W::Integer=L)
square_lattice(param::Dict)

Generates square lattice with size L=param["L"] $\times$ W=param["W"].

nsitetypes is 2: all 1 (2) sites do not connect to another 1 (2) site.

nbondtypes is 2: 1 bonds are parallel to x axis and 2 are parallel to y axis.

source

SpinMonteCarlo.triangular_lattice — Function.

triangular_lattice(L::Integer, W::Integer=L)
triangular_lattice(param::Dict)

Generates triangular lattice with size L=param["L"] $\times$ W=param["W"].

nsitetypes is 3: all 1, 2, and 3 sites do not connect to another 1, 2, and 3 site, respectively.

nbondtypes is 3: 1, 2, and 3 bonds make an angle of 0, 60, and 120 degree with x axis, respectively.

source

SpinMonteCarlo.cubic_lattice — Function.

cubic_lattice(L::Integer, W::Integer=L, H::Integer=W)
cubic_lattice(param::Dict)

Generates cubic lattice with size L=param["L"] $\times$ W=param["W"] $\times$ H=param["H"].

nsitetypes is 2: all 1 (2) sites do not connect to another 1 (2) site.

nbondtypes is 2: 1, 2, and 3 bonds are parallel to x, y, and z axis, respectively.

source

SpinMonteCarlo.fully_connected_lattice — Function.

fully_connected_lattice(N::Integer)
fully_connected_lattice(param::Dict)

Generates N=param["N"] site fully connected lattice

Both nsitetypes and nbondtypes are 1.

source

Update method

An index of model parameter (e.g., Js) is corresponding to sitetype or bondtype.

SpinMonteCarlo.local_update! — Function.

local_update!(model, param)
local_update!(model, T::Real, Js::AbstractArray)

Updates spin configuration by local spin flip and Metropolice algorithm under the temperature T = param["T"] and coupling constants J = param["J"]

source

SpinMonteCarlo.SW_update! — Function.

SW_update!(model, param::Parameter)
SW_update!(model, T::Real, Js::AbstractArray)

Updates spin configuration by Swendsen-Wang algorithm under temperature T=param["T"] and coupling constants J=param["J"]

source

SpinMonteCarlo.Wolff_update! — Function.

Wolff_update!(model, param::Parameter)
Wolff_update!(model, T::Real, Js::AbstractArray)

Updates spin configuration by Wolff algorithm under temperature T=param["T"] and coupling constants J=param["J"]

source

SpinMonteCarlo.loop_update! — Function.

loop_update!(model, param::Parameter)
loop_update!(model, T::Real,
             Jz::AbstractArray,
             Jxy::AbstractArray,
             Gamma:AbstractArray)

Updates spin configuration by loop algorithm under the temperature T = param["T"] and coupling constants Jz, Jxy and transverse field Gamma

source

Estimator

SpinMonteCarlo.simple_estimator — Function.

simple_estimator(model::Ising, T::Real, Js::AbstractArray)
simple_estimator(model::Potts, T::Real, Js::AbstractArray)

Returns the following observables as Dict{String, Any}

Observables

"Energy"

Energy per spin (site)

"Energy^2"

"Magnetization"

Total magnetization per spin (order paremeter)

"|Magnetization|"

"Magnetization^2"

"Magnetization^4"

source

simple_estimator(model::Clock, T::Real, Js::AbstractArray)
simple_estimator(model::XY, T::Real, Js::AbstractArray)

Returns the following observables as Dict{String, Any}

Observables

"Energy"

Energy per spin (site)

"Energy^2"

"|Magnetization|"

Absolute value of total magnetization per spin (order paremeter)

"|Magnetization|^2"

"|Magnetization|^4"

"Magnetization x"

x component of total magnetization per spin (order paremeter)

"|Magnetization x|"

"Magnetization x^2"

"Magnetization x^4"

"Magnetization y"

y component of total magnetization per spin (order paremeter)

"|Magnetization y|"

"Magnetization y^2"

"Magnetization y^4"

"Helicity Modulus x"

"Helicity Modulus y"

source

SpinMonteCarlo.improved_estimator — Function.

improved_estimator(model::Ising, T::Real, Js::AbstractArray, sw::SWInfo)

Returns the following observables as Dict{String, Any} using cluster information sw

Observables

"Energy"

Energy per spin (site)

"Energy^2"

"Magnetization"

Total magnetization per spin (site)

"|Magnetization|"

"|Magnetization|^2"

"|Magnetization|^4"

source

improved_estimator(model::Potts, T::Real, Js::AbstractArray, sw::SWInfo)

Returns the following observables as Dict{String, Any} using cluster information sw

Observables

"Energy"

"Energy^2"

"Magnetization"

"|Magnetization|"

"|Magnetization|^2"

"|Magnetization|^4"

source

improved_estimator(model::QuantumXXZ, T::Real, Js::AbstractArray, uf::UnionFind)

Returns the following observables as Dict{String, Any} using loop information uf

Observables

"Sign"

Sign of the weight function

"Sign * Energy"

Energy per spin (site)

"Sign * Energy^2"

"Sign * Magnetization"

Total magnetization (Sz) per spin (site)

"Sign * |Magnetization|"

"Sign * Magnetization^2"

"Sign * Magnetization^4"

source

Utility

SpinMonteCarlo.convert_parameter — Function.

convert_parameter(model, param)

Generates arguments of updater and estimator.

Example

julia> model = Ising(chain_lattice(4));

julia> p = convert_parameter(model, Parameter("J"=>1.0))
(1.0, [1.0, 1.0]) # T and Js

julia> p = convert_parameter(model, Parameter("J"=>[1.5, 0.5]))
(1.0, [1.5, 0.5]) # J can take a vector whose size is `numbondtypes(model)`

julia> model.spins
4-element Array{Int64,1}:
  1
  1
  1
 -1

julia> local_update!(model, p...);

julia> model.spins
4-element Array{Int64,1}:
  1
  1
 -1
 -1

source

SpinMonteCarlo.convert_parameter — Method.

convert_parameter(model::Union{Ising, Potts, Clock, XY}, param::Parameter)

Keynames:

"T": a scalar (default: 1.0).

"J": a vector with numbondtypes(model) elements (default: 1.0).

source

SpinMonteCarlo.convert_parameter — Method.

convert_parameter(model::QuantumXXZ, param::Parameter)

Keynames:

"T": a scalar (default: 1.0).

"Jz": a vector with numbondtypes(model) elements (default: 1.0).

"Jxy": a vector with numbondtypes(model) elements (default: 1.0).

"Gamma": a vector with numsitetypes(model) elements (default: 0.0).

source