Internals

Internal APIs

Documentation for SpinMonteCarlo.jl's internals.

Driver

SpinMonteCarlo.accumulateObservables! — Function.

accumulateObservables!(model, obs::MCObservableSet, localobs::Dict)

Accumulates localobs into obs. For example, obs["Energy"] << localobs["Energy"].

source

SpinMonteCarlo.postproc — Function.

postproc(model::Model, param::Dict, obs::MCObservableSet)

Post process of observables. For example, Specific heat will be calculated from energy, energy^2, and temperature.

source

Model

SpinMonteCarlo.LoopElementType — Type.

Enumtype including LET_*

source

SpinMonteCarlo.LET_Cut — Constant.

Loop element depicted as

|
o
|

or matrix

1 1
1 1

source

SpinMonteCarlo.LET_FMLink — Constant.

Loop element depicted as

|  |
|--|
|  |

or matrix

1 0 0 0
0 0 0 0
0 0 0 0
0 0 0 1

source

SpinMonteCarlo.LET_AFLink — Constant.

Loop element depicted as

|  |
|~~|
|  |

or matrix

0 0 0 0
0 1 0 0
0 0 1 0
0 0 0 0

source

SpinMonteCarlo.LET_Vertex — Constant.

Loop element depicted as

|  |
~~~~
~~~~
|  |

or matrix

0 0 0 0
0 1 1 0
0 1 1 0
0 0 0 0

source

SpinMonteCarlo.LET_Cross — Constant.

Loop element depicted as

|  |
 \/
 /\
|  |

or matrix

1 0 0 0
0 0 1 0
0 1 0 0
0 0 0 1

source

SpinMonteCarlo.LocalLoopOperator — Type.

(Imaginary-temporary and spatial) local operator as a perturbation with assigned loop element.

Fields

let_type : assigned loop element

isdiagonal : operator is diagonal or not

in other words, two states connecting this perturbation are equivalent to each other or not.

time : imaginary time ($\tau/\beta \in [0,1)$) which this perturbation acts on.

space : index of subspin or subbond which this perturbation acts on.

bottom_id :: index of node of union find assigned to a loop

top_id :: index of node of union find assigned to the other loop

source

Utility

SpinMonteCarlo.default_estimator — Function.

default_estimator(model, updatemethod!)

Determines estimator to be used when param["Estimator"] is not set.

source

SpinMonteCarlo.@gen_convert_parameter — Macro.

@gen_convert_parameter(model_typename, (keyname, size_fn, default)...)

Generates convert_parameter(model::model_typename, param::Parameter).

Example

@gen_convert_parameter(A, ("A", numbondtypes, 1.0), ("B", 1, 1))

generates a function equivalent to the following:

doc"""
    convert_parameter(model::A, param::Parameter)

# Keynames

- "A": a vector with `numbondtypes(model)` elements (default: 1)
- "B": a scalar (default: 1.0)

"""
function convert_parameter(model::A, param::Parameter)
    ## if `size_fn` is a `Function`,
    ## result is a vector whose size is `size_fn(model)`.
    ## `param["A"]` can take a scalar or a vector.
    a = get(param, "A", 1.0)
    as = zeros(Float64, numbondtypes(model))
    as .= a

    ## otherwise,
    ## result is a scalar.
    b = get(param, "B", 1) :: Int

    return as, b
end

source

SpinMonteCarlo.UnionFind — Type.

Union-find algorithm.

source

SpinMonteCarlo.addnode! — Function.

addnode!(u::UnionFind)

Adds a new node into u and returns the number of nodes including the added node.

source

SpinMonteCarlo.unify! — Function.

unify!(u, n1, n2)

Connects n1 and n2 nodes and returns the root.

source

SpinMonteCarlo.clusterize! — Function.

clusterize!(u::UnionFind)

Assigns cluster ID to each node and returns the number of clusters.

source

SpinMonteCarlo.clusterid — Function.

clusterid(u::UnionFind, i::Integer)

Returns the index of the cluster where i node belongs.

source